# Speed up counting past 30 day traffic via ingest_counters