--- a/priv/ingest_repo/migrations/20241120064325_create_ingest_counters_site_traffic_projection.exs
+++ b/priv/ingest_repo/migrations/20241120064325_create_ingest_counters_site_traffic_projection.exs
@@ -2,6 +2,20 @@ defmodule Plausible.IngestRepo.Migrations.CreateIngestCountersSiteTrafficProject
   use Ecto.Migration
 
   def up do
+    # ClickHouse 24.8 introduced `deduplicate_merge_projection_mode` which we need to set to create this projection.
+    # As such, try to set the setting and if it fails, ignore it.
+    execute(fn ->
+      try do
+        repo().query!("""
+          ALTER TABLE ingest_counters
+          #{Plausible.MigrationUtils.on_cluster_statement("ingest_counters")}
+          MODIFY SETTING deduplicate_merge_projection_mode='rebuild'
+        """)
+      rescue
+        _e -> nil
+      end
+    end)
+
     execute """
       ALTER TABLE ingest_counters
       #{Plausible.MigrationUtils.on_cluster_statement("ingest_counters")}
@@ -13,10 +27,6 @@ defmodule Plausible.IngestRepo.Migrations.CreateIngestCountersSiteTrafficProject
   end
 
   def down do
-    execute """
-      ALTER TABLE ingest_counters
-      #{Plausible.MigrationUtils.on_cluster_statement("ingest_counters")}
-      DROP PROJECTION ingest_counters_site_traffic_projection
-    """
+    raise "irreversible"
   end
 end