Companion to https://github.com/plausible/analytics/pull/4839

A projection speeds up ClickHouse queries by storing data in a different order in each part. In this case, it pre-aggregates results by site, speeding up the query and requiring less resources at query-time.

Query stats before:
```
-- 10 rows in set. Elapsed: 1.134 sec. Processed 1.45 billion rows, 21.65 GB (1.28 billion rows/s., 19.09 GB/s.)
-- Peak memory usage: 53.41 MiB.
```

After:
```
-- 10 rows in set. Elapsed: 0.321 sec. Processed 39.70 million rows, 1.07 GB (123.51 million rows/s., 3.33 GB/s.)
-- Peak memory usage: 97.43 MiB.
```